Updated Dockerfile to fix compilation errors and README for missing command in docker setup
qt5-defualt, which installs various dependencies of qt5 for Linux is not available on latest version of Ubuntu. To allow docker file to build the image, the dependencies included in the qt5-default were manually added to the dockerfile. Readme was also updated to add missing command in the build image step for seting up Docker enviroment.